+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.9.0
+- feat: Add libp2p-bitswap and feature gate both implementation. [PR 126](https://github.com/dariusc93/rust-ipfs/pull/126)
 - feat: Add `Repo::pin`, `Repo::remove_pin` and expanded options [PR 122](https://github.com/dariusc93/rust-ipfs/pull/122)
 - chore: Update `rust-ipns` [PR 121](https://github.com/dariusc93/rust-ipfs/pull/121)